Only define the proxy fixture in test_proxy
